ao486: extract CD code to separate file, some refactoring and fixes.
