+++ MiSTer.vcxproj.filters
@@ -205,6 +205,9 @@
     <ClCompile Include="support\x86\x86_ide.cpp">
       <Filter>Source Files\support</Filter>
     </ClCompile>
+    <ClCompile Include="support\x86\x86_cdrom.cpp">
+      <Filter>Source Files\support</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="battery.h">
@@ -388,6 +391,9 @@
       <Filter>Header Files\support</Filter>
     </ClInclude>
     <ClInclude Include="support\x86\x86_ide.h">
+      <Filter>Header Files\support</Filter>
+    </ClInclude>
+    <ClInclude Include="support\x86\x86_cdrom.h">
       <Filter>Header Files\support</Filter>
     </ClInclude>
   </ItemGroup>
